Prelimilary Smolproc FPGA Implementation
This implementation of Smolproc has existed since the video on my channel showcasing Smolproc running fibonacci on an FPGA was uploaded but I never got around to uploading it until now. This is the raw project directory and my code is frankly awful, but then again this was my first FPGA project.